Move InvokeResult/Error to types.rs and derive Clone
