+++ concordium-std/src/types.rs
@@ -15,7 +15,7 @@ pub struct ExternParameter {
 
 /// A type representing the return value of contract calls.
 /// This is when a contract calls another contract, it may get a return value.
-#[derive(Debug)]
+#[derive(Debug, Clone)]
 pub struct CallResponse {
     /// The index of the call response.
     pub(crate) i:                NonZeroU32,
@@ -26,6 +26,35 @@ pub struct CallResponse {
 pub struct ReturnValue {
     pub(crate) current_position: u32,
 }
+
+/// FIXME: Have two error types, one for transfers, one for calls.
+/// FIXME: Consider adding `#[non_exhaustive]`.
+#[repr(i32)]
+#[derive(Debug, Clone)]
+pub enum InvokeError {
+    /// Amount that was to be transferred is not available to the sender.
+    AmountTooLarge,
+    /// Account that is to be transferred to does not exist.
+    MissingAccount,
+    /// Contract that is to be transferred to does not exist.
+    MissingContract,
+    /// The contract to be invoked exists, but the entrypoint that was named
+    /// does not.
+    MissingEntrypoint,
+    /// Sending a message to the V0 contract failed.
+    MessageFailed,
+    /// Contract that was called rejected with the given reason.
+    LogicReject {
+        reason:       i32,
+        return_value: CallResponse,
+    },
+    /// Execution of a contract call triggered a runtime error.
+    Trap,
+    /// Unrecognized error. Reser
+    Unknown,
+}
+
+pub type InvokeResult<A> = Result<A, InvokeError>;
 
 /// A type representing the attributes, lazily acquired from the host.
 #[derive(Default)]
